kernel: Simplify syscall sethostname
Currently syscall function definition for 'sethostname' is a candidate
for refactoring.  The uts_sem semaphore is held while a copy_from_user()
call is made, it is unnecessary to hold the semaphore during this call.
Keeping with locking principles locks should be held for the minimum
amount of time necessary.  An extra local variable 'errno' is currently
used so as to be able to return the correct error code after the
semaphore is released (up'ed).  We can simplify the code by doing the
copy_from_user() before acquiring the semaphore.

Do copy_from_user() before acquiring 'uts_sem' semaphore.  Return
error (-EFAULT) directly if copy_from_user() fails.

Signed-off-by: Tobin C. Harding <me@tobin.cc>