kernel: Use variable name 'nbytes' instead of 'i'
Currently code uses the variable name 'i' for the number of bytes to be
copied to the user, this is _not_ a descriptive variable name and also
violates the principle of using one name for one thing ('i' is typically used as
a loop counter variable name).  We can use the variable name 'nbytes' to
better describe this variable thereby reducing cognitive load for future
developers reading the code.  While we are touching these lines and
making the code easier to read we can put whitespace around the
ups/downs to the semaphore.  Again this (slightly) reduces cognitive load
since the up/down pair are more easily seen.

Signed-off-by: Tobin C. Harding <me@tobin.cc>