slub: Add defrag_used_ratio field and sysfs support
In preparation for enabling defragmentation of slab pages.

"defrag_used_ratio" is used to set the threshold at which
defragmentation should be attempted on a slab page.

"defrag_used_ratio" is a percentage in the range of 0 - 100 (inclusive).
If less than that percentage of slots in a slab page are in use then the
slab page will become subject to defragmentation.

Add a defrag ratio field and set it to 30% by default. A limit of 30%
specifies that more than 3 out of 10 available slots for objects need to
be in use otherwise slab defragmentation will be attempted on the
remaining objects.

Co-developed-by: Christoph Lameter <cl@linux.com>
Signed-off-by: Tobin C. Harding <tobin@kernel.org>