slub: Enable slab defragmentation using SMO
If many objects are allocated with the slab allocator and freed in an
arbitrary order then the slab caches can become internally fragmented.
Now that the slab allocator supports movable objects we can defragment
any cache that has this feature enabled.

Slab defragmentation may occur:

1. Unconditionally when __kmem_cache_shrink() is called on a slab cache
   by the kernel calling kmem_cache_shrink().

2. Unconditionally through the use of the slabinfo command.

	slabinfo <cache> -s

3. Conditionally via the use of kmem_cache_defrag()

Use SMO when shrinking cache.  Currently when the kernel calls
kmem_cache_shrink() we curate the partial slabs list.  If object
migration is not enabled for the cache we still do this, if however SMO
is enabled, we attempt to move objects in partially full slabs in order
to defragment the cache.  Shrink attempts to move all objects in order
to reduce the cache to a single partial slab for each node.

kmem_cache_defrag() differs from shrink in that it operates dependent on
the defrag_used_ratio and only attempts to move objects if the number of
partial slabs exceeds MAX_PARTIAL (for each node).

Add function kmem_cache_defrag(int node).

   kmem_cache_defrag() only performs defragmentation if the usage ratio
   of the slab is lower than the configured percentage (sysfs file added
   in previous patch).  Fragmentation ratios are measured by calculating
   the percentage of objects in use compared to the total number of
   objects that the slab page can accommodate.

   The scanning of slab caches is optimized because the defragmentable
   slabs come first on the list. Thus we can terminate scans on the
   first slab encountered that does not support defragmentation.

   kmem_cache_defrag() takes a node parameter. This can either be -1 if
   defragmentation should be performed on all nodes, or a node number.

   Defragmentation may be disabled by setting defrag ratio to 0

	echo 0 > /sys/kernel/slab/<cache>/defrag_used_ratio

In order for a cache to be defragmentable the cache must support object
migration (SMO).  Enabling SMO for a cache is done via a call to the
recently added function:

	void kmem_cache_setup_mobility(struct kmem_cache *,
				       kmem_cache_isolate_func,
			               kmem_cache_migrate_func);

Co-developed-by: Christoph Lameter <cl@linux.com>
Signed-off-by: Tobin C. Harding <tobin@kernel.org>