fs: Guard unusual text with backticks
Currently we have unusual text which in context is clear

	@flags: O_... flags with which the new file will be opened

Currently this docstring is _not_ included in any documentation file but
if it is in future this will cause Sphinx to emit a warning (as
discovered while documenting alloc_file_pseudo()).  Let's fix it now to
save our future selves the hassle.

Guard unusual text with backticks.

Signed-off-by: Tobin C. Harding <tobin@kernel.org>