dcache: Clean up function docstring members
We just cleaned up the function docstrings, we omitted while doing that
clean up the member descriptions.  This was done to ease the cognitive
load required for review.  Seeing what has changed in a line with an
added period can be taxing on the patch reader, if we do all these
changes as a separate patch its easier to review.

Clean up function docstring member descriptions:n
 - Capitalise the initial character.
 - Add period (full stop) to end of string.
 - iff required flesh out the description.

Signed-off-by: Tobin C. Harding <tobin@kernel.org>